refactored out partials for rant console, search input, favorites along with helper methods

+++ app/helpers/users_helper.rb
@@ -30,15 +30,15 @@ module UsersHelper
   end
 
   def dash_favorite_button rant
-    button_to 'Favorite',
-              rant_favorites_path(rant.id),
-              {rant_id: rant.id}
-  end
-
-  def dash_unfavorite_button rant
-    button_to 'Unfavorite',
-              rant_favorite_path(id: current_user.id, rant_id: rant.id),
-              method: :delete
+    if favorite_rant(rant)
+      button_to 'Unfavorite',
+                rant_favorite_path(id: current_user.id, rant_id: rant.id),
+                method: :delete
+    else
+      button_to 'Favorite',
+                rant_favorites_path(rant.id),
+                {rant_id: rant.id}
+    end
   end
 
   def favorite_rant rant
